temp_enrichment/requirements.txt
# Минимальные зависимости для temp_enrichment модуля
# Совместимо с основным проектом rag-stroyactiv-pyton

numpy>=1.20.0
openai>=1.12.0,<2.0.0  # Поддержка legacy и новых версий
packaging>=23.0
pyyaml>=6.0
python-dotenv>=0.19.0  # Для загрузки переменных окружения

# Опциональные зависимости (если не установлены в основном проекте)
# pydantic>=2.0.0
# fastapi>=0.100.0 